kmk/output.c: signed compare warning.
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@3243 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kmk/output.c
@@ -511,7 +511,7 @@ output_write_text (struct output *out, int is_err, const char *src, size_t len)
       while (1)
         {
           EINTRLOOP (r, write (fd, src, len));
-          if (r == len || r <= 0)
+          if ((size_t)r == len || r <= 0)
             break;
           len -= r;
           src += r;
